Extract CLI exit code taxonomy into reusable microcrate
