+++ Cargo.toml
@@ -28,6 +28,7 @@ members = [
     "crates/copybook-rdw-predicates",
     "crates/copybook-safe-index",
     "crates/copybook-safe-text",
+    "crates/copybook-exit-codes",
     "crates/copybook-cli",
     "crates/copybook-cli-determinism",
     "crates/copybook-governance",
@@ -75,6 +76,7 @@ copybook-overpunch = { version = "=0.4.3", path = "crates/copybook-overpunch" }
 copybook-determinism = { version = "=0.4.3", path = "crates/copybook-determinism" }
 copybook-safe-ops = { version = "=0.4.3", path = "crates/copybook-safe-ops" }
 copybook-safe-text = { version = "=0.4.3", path = "crates/copybook-safe-text" }
+copybook-exit-codes = { version = "=0.4.3", path = "crates/copybook-exit-codes" }
 copybook-safe-index = { version = "=0.4.3", path = "crates/copybook-safe-index" }
 copybook-rdw-predicates = { version = "=0.4.3", path = "crates/copybook-rdw-predicates" }
 copybook-codepage = { version = "=0.4.3", path = "crates/copybook-codepage" }
