+++ crates/copybook-cli-determinism/Cargo.toml
@@ -17,6 +17,7 @@ anyhow.workspace = true
 clap.workspace = true
 serde_json.workspace = true
 copybook-determinism = { workspace = true }
+copybook-exit-codes = { workspace = true }
 copybook-core = { workspace = true }
 copybook-codec = { workspace = true }
