+++ crates/copybook-cli/Cargo.toml
@@ -29,6 +29,7 @@ path = "src/main.rs"
 [dependencies]
 clap.workspace = true
 copybook-cli-determinism = { workspace = true }
+copybook-exit-codes = { workspace = true }
 serde_json.workspace = true
 serde_plain.workspace = true
 tracing.workspace = true
