+++ crates/copybook-exit-codes/Cargo.toml
@@ -1,0 +1,26 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+[package]
+name = "copybook-exit-codes"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Canonical CLI exit-code taxonomy and serialization for copybook-rs"
+keywords.workspace = true
+categories.workspace = true
+
+[lib]
+name = "copybook_exit_codes"
+path = "src/lib.rs"
+
+[dependencies]
+serde.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
+
+[lints]
+workspace = true
